+++ apps/demo/proxy.ts
@@ -9,8 +9,8 @@ const localeMap = {
 
 export const directusConfig: TDirectusRouteConfig<keyof typeof localeMap> = {
   localeMap,
-  localePrefix: 'as-needed',
-  defaultLocale: 'fr-CA',
+  localePrefix: 'always',
+  defaultLocale: 'en-CA',
   collectionSettings: {
     default: {
       idField: 'id',
@@ -23,5 +23,5 @@ export async function proxy(request: NextRequest) {
 }
 
 export const config = {
-  matcher: ['/((?!api|_next/static|_next/image|blocks/|favicon.ico|react_devtools_backend_compact.js.map).*)'],
+  matcher: ['/((?!api|_next/static|_next/image|blocks/|favicon.ico|react_devtools_backend_compact.js.map).*)', '/directus-next/img-img-loader', '/directus-next/imgix-image-loader'],
 }