+++ libs/directus/directus-next/src/index.ts
@@ -1,5 +1,7 @@
 export * from './draft'
 export type { TFiles } from './files/interface'
+export { default as directusImageLoader } from './loaders/directusImageLoader'
+export { default as imgixImageLoader } from './loaders/imgixImageLoader'
 export { logger as DirectusNextLogger } from './logger'
 export * from './redirect'
 export { getJsonErrorResponse } from './response'