+++ libs/directus/directus-block/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.1.5 (2026-07-13)
+
+### 🩹 Fixes
+
+- **directus-block:** mergeConfigs accepts type generic parameters ([#517](https://github.com/OKAMca/stack/pull/517))
+
+### ❤️ Thank You
+
+- Pierre-Olivier Clerson @poclerson
+
 ## 2.1.4 (2026-04-23)
 
 ### 🧱 Updated Dependencies