+++ libs/directus/directus-next/CHANGELOG.md
@@ -1,3 +1,37 @@
+## 2.3.1 (2026-07-15)
+
+### 🚀 Features
+
+- **directus-next:** imgix & directus image loaders ([#463](https://github.com/OKAMca/stack/pull/463))
+
+### 🩹 Fixes
+
+- **next-component:** only export localeprefix from index.ts barrel ([#522](https://github.com/OKAMca/stack/pull/522))
+
+### ❤️ Thank You
+
+- Pierre-Olivier Clerson @poclerson
+
+## 2.3.0 (2026-07-14)
+
+### 🚀 Features
+
+- **directus-next:** imgix & directus image loaders ([#463](https://github.com/OKAMca/stack/pull/463))
+
+### ❤️ Thank You
+
+- Pierre-Olivier Clerson @poclerson
+
+## 2.2.6 (2026-07-13)
+
+This was a version bump only for directus-next to align it with other projects, there were no code changes.
+
+## 2.2.5 (2026-04-27)
+
+### 🧱 Updated Dependencies
+
+- Updated next-component to 2.1.5
+
 ## 2.2.3 (2026-04-15)
 
 ### 🧱 Updated Dependencies